Fix an edge case
If a track end is exactly on the edge of a via, it can count as "inside" the via by IsPointOnEnds() but "outside" the via by __PointDistance(). It seems more reliable overall to check if the start of a track is outside, rather than if the end of a track is inside.